+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
-## [2.6.0.canary.2]
+## [Unreleased]
 
 ### Added
+
+- `throttle` in common utils
+- `debounce` in common utils
+
+## [2.6.0.canary.2]
 
 - `getScrollbarWidth` options in browser utils
 - `disableElementScroll` in browser utils